backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
pillow==11.3.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
mediapipe==0.10.21
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.35.1
pytest==8.4.1
httpx==0.28.1
